requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
Django==5.2.5
django-cloudinary-storage==0.3.0
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
geographiclib==2.0
geopy==2.4.1
idna==3.10
mollie-api-python==3.8.0
oauthlib==3.3.1
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
